# backend/openapi.yaml
openapi: 3.0.3
info:
  title: EduVerify API
  version: "1.0.0"
  description: >
    API REST del backend de EduVerify. Todas las respuestas siguen el envelope
    { status: 'success', data } / { status: 'error', message }.
    Ver backend/README.md para setup y arquitectura.
servers:
  - url: http://localhost:3001
security: []
...
paths:
...
  /api/users/{id}/videos:
    get:
      tags: [users]
      summary: Videos publicados por un usuario (paginado)
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VideoListResponse' }

  # ── VIDEOS ────────────────────────────────────────────────────────────────
  /api/videos:
    get:
      tags: [videos]
      summary: Listar videos (paginado, filtro opcional por categoría)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: categoria
          in: query
          schema: { $ref: '#/components/schemas/Categoria' }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VideoListResponse' }
    post:
      tags: [videos]
      summary: Crear video (solo profesor/creador)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [titulo, url_video, categoria]
              properties:
                titulo: { type: string, minLength: 3, maxLength: 255 }
                descripcion: { type: string }
                url_video: { type: string, format: uri }
                categoria: { $ref: '#/components/schemas/Categoria' }
                tipo: { type: string, enum: [grabado, envivo], default: grabado }
                es_premium: { type: boolean, default: false }
                duracion: { type: string }
      responses:
        "201":
          description: Video creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Video' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /api/videos/{id}:
    get:
      tags: [videos]
      summary: Detalle de un video
      parameters:
        - $ref: '#/components/parameters/VideoId'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Video' }
        "404": { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [videos]
      summary: Editar video (solo el autor)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/VideoId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                titulo: { type: string, minLength: 3, maxLength: 255 }
                descripcion: { type: string }
                categoria: { $ref: '#/components/schemas/Categoria' }
                es_premium: { type: boolean }
                duracion: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Video' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403":
          description: No es el autor del video
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [videos]
      summary: Eliminar video (solo el autor)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/VideoId'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403":
          description: No es el autor del video
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/videos/{id}/view:
    post:
      tags: [videos]
      summary: Incrementar contador de vistas
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/VideoId'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      vistas: { type: integer }
        "401": { $ref: '#/components/responses/Unauthorized' }

  # ── COMMENTS ──────────────────────────────────────────────────────────────
...
  /api/playlists/{id}/videos/{videoId}:
    post:
      tags: [playlists]
      summary: Añadir video a playlist (upsert, solo el dueño)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: videoId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [playlists]
      summary: Quitar video de playlist (solo el dueño)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: videoId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── PROFESOR PLAYLISTS ───────────────────────────────────────────────────
...
  /api/profesor/playlists/{id}/videos/{videoId}:
    post:
      tags: [profesor-playlists]
      summary: Añadir/reordenar video en playlist (upsert `orden`, solo el dueño)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: videoId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                orden: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [profesor-playlists]
      summary: Quitar video de playlist (solo el dueño)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: videoId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── SUBSCRIPTIONS ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token obtenido de /api/auth/login, /registro o /google. Payload:
        { sub: userId, email, rol, premium }. Expira en 7 días.
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 50 }
    UserId:
      name: id
      in: path
      required: true
      schema: { type: integer }
    VideoId:
      name: id
      in: path
      required: true
      schema: { type: integer }
  responses:
    BadRequest:
      description: Body/params inválidos (Zod)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Token ausente, inválido o expirado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Rol insuficiente o no es el dueño del recurso
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Categoria:
      type: string
      enum: [Programación, Ciberseguridad, Matemáticas, Electrónica, Arte]
    ErrorResponse:
      type: object
      properties:
        status: { type: string, example: error }
        message: { type: string }
    SuccessMessage:
      type: object
      properties:
        status: { type: string, example: success }
        data:
          type: object
          properties:
            message: { type: string }
...
    Video:
      type: object
      properties:
        id: { type: integer }
        titulo: { type: string }
        descripcion: { type: string, nullable: true }
        url_video: { type: string, format: uri }
        categoria: { $ref: '#/components/schemas/Categoria' }
        tipo: { type: string, enum: [grabado, envivo] }
        es_premium: { type: boolean }
        vistas: { type: integer }
        duracion: { type: string }
        created_at: { type: string, format: date-time }
        autor: { type: string }
        autor_id: { type: integer }
        author_avatar_url: { type: string, format: uri, nullable: true }
    VideoListResponse:
      type: object
      properties:
        status: { type: string, example: success }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Video' }
            total: { type: integer }
            page: { type: integer }
            limit: { type: integer }
...